Conflicts:
	DM2231_Lab2_MVC/DM2231_Lab2/Bullet.cpp
	DM2231_Lab2_MVC/DM2231_Lab2/Bullet.h
	DM2231_Lab2_MVC/DM2231_Lab2/Collision.cpp
	DM2231_Lab2_MVC/DM2231_Lab2/Collision.h
	DM2231_Lab2_MVC/DM2231_Lab2/DM2231_Controller.cpp
	DM2231_Lab2_MVC/DM2231_Lab2/DM2231_Lab2.vcxproj
	DM2231_Lab2_MVC/DM2231_Lab2/DM2231_Lab2.vcxproj.filters
	DM2231_Lab2_MVC/DM2231_Lab2/DM2231_Model.cpp
	DM2231_Lab2_MVC/DM2231_Lab2/DM2231_Model.h
	DM2231_Lab2_MVC/DM2231_Lab2/Entity.cpp
	DM2231_Lab2_MVC/DM2231_Lab2/Entity.h
	DM2231_Lab2_MVC/DM2231_Lab2/EntityFactory.cpp
	DM2231_Lab2_MVC/DM2231_Lab2/EntityFactory.h
	DM2231_Lab2_MVC/DM2231_Lab2/PlayerInfo.h

+++ DM2231_Lab2_MVC/DM2231_Lab2/Mouse.h
@@ -41,7 +41,7 @@ public:
 			this->last_x = x;
 			this->last_y = y;
 		}
-		MousePos.Set(x,y);
+		MousePos.Set(static_cast<float>(x),static_cast<float>(y));
 		this->x = new_x;
 		this->y = new_y;
 	}
